+++ Cafe_App/Data/Masa.cs
@@ -15,9 +15,9 @@ public partial class Masa
 
     public bool Temizlik { get; set; }
 
-    public int Tutar { get; set; }
+    public decimal Tutar { get; set; }
 
-    public int OdenenTutar { get; set; }
+    public decimal OdenenTutar { get; set; }
 
     public int PersonelId { get; set; }
